% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdpclust.R
\name{bigdpclust}
\alias{bigdpclust}
\title{Gaussian Dirichlet Process Mixture CLustering For Tall Data}
\usage{
bigdpclust(data, coresets = NULL, clumping_fn = stats::kmeans,
  nclumps = min(500, nrow(data)/10), hyperG0 = NULL, Ninit = 50,
  Nmcmc = 1000, burnin = Nmcmc/5, thin = 2, loss_fn = "MBinderN",
  diagVar = FALSE, plotevery_nit = Nmcmc/10)
}
\arguments{
\item{data}{n by p matrix wuith n observations in rows and p dimensions in columns.}

\item{coresets}{a list with 3 components}

\item{burnin}{an integer giving the number of MCMC iterations to burn. Ddefaults is half)}

\item{plotevery_nit}{an integer indicating the interval between plotted iterations
when \code{doPlot} is \code{TRUE}. Default is \code{Nmcmc/10}}

\item{variance}{logical flag indicating whether the covariance matrix of each cluster is
constrained as diagonal, or unconstrained full matrix.
Default is \code{FALSE} (unconstrained covariance).}
}
\description{
Gaussian Dirichlet Process Mixture CLustering For Tall Data
}
\examples{
n1 <- 50000
n2 <- 500
mydata <- rbind(cbind(rnorm(n1), rnorm(n = n1)),
                cbind(rnorm(n2, m=10), rnorm(n = n2, m=10)))
#plot(mydata)
#coresets <- stats::kmeans(mydata, centers = 100)[c("cluster", "centers", "size")]

res <- bigdpclust(mydata, nclumps=200)
table(res$cluster[1:n1])
table(res$cluster[n1 + 1:n2])
}
\author{
Boris Hejblum, Paul Kirk
}
