% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigdpclust.R
\name{bigdpclust}
\alias{bigdpclust}
\title{Gaussian Dirichlet Process Mixture CLustering For Tall Data}
\usage{
bigdpclust()
}
\description{
Gaussian Dirichlet Process Mixture CLustering For Tall Data
}
\author{
Boris Hejblum, Paul Kirk
}
